workaround for Alpha - by Falk Hueffner <falk.hueffner@student.uni-tuebingen.de>
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@3537 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ loader/wine/vfw.h
@@ -19,17 +19,17 @@ typedef struct __attribute__((__packed__))
 #define _BITMAPINFOHEADER_
 typedef struct __attribute__((__packed__))
 {
-    long 	biSize;
-    long  	biWidth;
-    long  	biHeight;
+    int 	biSize;
+    int  	biWidth;
+    int  	biHeight;
     short 	biPlanes;
     short 	biBitCount;
-    long 	biCompression;
-    long 	biSizeImage;
-    long  	biXPelsPerMeter;
-    long  	biYPelsPerMeter;
-    long 	biClrUsed;
-    long 	biClrImportant;
+    int 	biCompression;
+    int 	biSizeImage;
+    int  	biXPelsPerMeter;
+    int  	biYPelsPerMeter;
+    int 	biClrUsed;
+    int 	biClrImportant;
 } BITMAPINFOHEADER, *PBITMAPINFOHEADER, *LPBITMAPINFOHEADER;
 typedef struct {
 	BITMAPINFOHEADER bmiHeader;
